requirements.txt
h5py==3.6.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy
opencv-python==4.5.5.64
packaging==21.3
pandas
Pillow==9.1.0
tensorflow==2.1.0
urllib3==1.26.9
